pyproject.toml
[tool.poetry]
name = "dla-ideal-solver"
version = "0.0.1"
description = "Diffusion-Limited Aggregation (DLA) solver with Numba acceleration"
authors = [
    "Sandy H. S. Herho <sandy.herho@email.ucr.edu>",
    "Siti N. Kaban",
    "Nurjanna J. Trilaksono",
    "Rusmawan Suwarman",
    "Dasapta E. Irawan"
]
license = "MIT"
readme = "README.md"
packages = [{include = "dla_ideal", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
numpy = ">=1.20.0"
matplotlib = ">=3.3.0"
netCDF4 = ">=1.5.0"
tqdm = ">=4.60.0"
numba = ">=0.53.0"
Pillow = ">=8.0.0"

[tool.poetry.group.dev.dependencies]
black = ">=22.0.0"

[tool.poetry.scripts]
dla-simulate = "dla_ideal.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
